requirements.txt
typer
httpx
attrs
pyyaml
